Move more stuff into files folder
