+++ apps/desktop/src/lib/files/filetree.test.ts
